+++ packages/app-store/metamask/README.mdx
@@ -1,7 +1,7 @@
 ---
 items:
-  - /api/app-store/metamaskweb3/example1.png
-  - /api/app-store/metamaskweb3/example2.png
+  - /api/app-store/metamask/example1.png
+  - /api/app-store/metamask/example2.png
 ---
 
 <Slider items={items} />
