+++ packages/app-store/metamask/api/add.ts
@@ -8,13 +8,11 @@ import prisma from "@calcom/prisma";
  * @param res
  */
 export default async function handler(req: NextApiRequest, res: NextApiResponse) {
-  console.log("🚀 ~ file: add.ts ~ line 11 ~ handler ~ req", req.method);
   if (!req.session?.user?.id) {
     return res.status(401).json({ message: "You must be logged in to do this" });
   }
 
   const appType = "metamask_web3";
-  console.log("🚀 ~ file: add.ts ~ line 17 ~ handler ~ appType", appType);
   try {
     const alreadyInstalled = await prisma.credential.findFirst({
       where: {
@@ -22,7 +20,6 @@ export default async function handler(req: NextApiRequest, res: NextApiResponse)
         userId: req.session.user.id,
       },
     });
-    console.log("🚀 ~ file: add.ts ~ line 25 ~ handler ~ alreadyInstalled", alreadyInstalled);
     if (alreadyInstalled) {
       throw new Error("Already installed");
     }
@@ -31,9 +28,9 @@ export default async function handler(req: NextApiRequest, res: NextApiResponse)
         type: appType,
         key: { isWeb3Active: true },
         userId: req.session.user.id,
+        appId: "metamask",
       },
     });
-    console.log("🚀 ~ file: add.ts ~ line 35 ~ handler ~ installation", installation);
     if (!installation) {
       throw new Error("Unable to create user credential for metamask");
     }
