+++ packages/app-store/metamask/components/InstallAppButton.tsx
@@ -2,7 +2,8 @@ import useAddAppMutation from "../../_utils/useAddAppMutation";
 import { InstallAppButtonProps } from "../../types";
 
 export default function InstallAppButton(props: InstallAppButtonProps) {
-  const mutation = useAddAppMutation("metamask_web3");
+  // @ts-ignore TODO: deprecate App types in favor of DB slugs
+  const mutation = useAddAppMutation("metamask");
 
   return (
     <>
