+++ packages/app-store/metamask/static/example1.png
