+++ packages/app-store/metamask/static/example2.png
