+++ packages/prisma/seed-app-store.ts
@@ -77,6 +77,7 @@ async function main() {
   }
   // Web3 apps
   await createApp("huddle01", "huddle01video", ["web3", "video"]);
+  await createApp("metamask", "metamask", ["web3"]);
   // Messaging apps
   if (process.env.SLACK_CLIENT_ID && process.env.SLACK_CLIENT_SECRET && process.env.SLACK_SIGNING_SECRET) {
     await createApp("slack", "slackmessaging", ["messaging"], {
